04_UART_PyQt5_Thread/requirements.txt
pyserial
pyqt5
requests